[agent] refactor(cli): extract io module from main.rs
