+++ crates/gaze-cli/src/io.rs
@@ -1,0 +1,48 @@
+use std::io::{self, Read};
+
+use crate::error::CliError;
+
+/// Default max-bytes cap for stdin. Keeps a runaway or attacker-controlled
+/// upstream from OOM'ing the worker. Override with `--max-bytes`.
+pub(crate) const DEFAULT_MAX_BYTES: u64 = 10 * 1024 * 1024;
+
+/// Read stdin up to `max_bytes + 1` and return the bytes.
+///
+/// Reading one extra byte past the cap lets us distinguish "input exactly
+/// at the limit" from "input exceeds the limit" without a second probe.
+pub(crate) fn read_stdin_bytes(max_bytes: u64) -> std::result::Result<Vec<u8>, CliError> {
+    let mut buf = Vec::new();
+    let limit = max_bytes.saturating_add(1);
+    io::stdin()
+        .take(limit)
+        .read_to_end(&mut buf)
+        .map_err(|_| CliError::Io)?;
+    if buf.len() as u64 > max_bytes {
+        return Err(CliError::InputTooLarge);
+    }
+    Ok(buf)
+}
+
+/// Read stdin as UTF-8 text, enforcing the size cap. Distinguishes:
+///   - 0 bytes              -> `EmptyInput`     (exit 1)
+///   - > max_bytes          -> `InputTooLarge` (exit 1)
+///   - non-UTF-8            -> `InvalidEncoding` (exit 1)
+///   - IO / OS error        -> `Io`            (exit 4)
+///
+/// `clean` calls this; `restore` uses the bytes path directly since the
+/// restore stdin is JSON and serde_json does its own UTF-8 validation.
+pub(crate) fn read_stdin_text(max_bytes: u64) -> std::result::Result<String, CliError> {
+    let bytes = read_stdin_bytes(max_bytes)?;
+    if bytes.is_empty() {
+        return Err(CliError::EmptyInput);
+    }
+    String::from_utf8(bytes).map_err(|_| CliError::InvalidEncoding)
+}
+
+pub(crate) fn require_json_format(format: &str) -> std::result::Result<(), CliError> {
+    if format == "json" {
+        Ok(())
+    } else {
+        Err(CliError::PolicyConfig)
+    }
+}
